+++ src/main/kotlin/io/kjson/ktor/JSONKtorFunctions.kt
@@ -26,25 +26,17 @@
 package io.kjson.ktor
 
 import kotlin.reflect.KType
-import kotlin.reflect.typeOf
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.flow.Flow
 
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.HttpRequestBuilder
-import io.ktor.client.request.prepareRequest
-import io.ktor.client.request.setBody
-import io.ktor.client.utils.EmptyContent
 import io.ktor.http.ContentType
 import io.ktor.http.HeaderValue
 import io.ktor.http.Headers
 import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
-import io.ktor.http.Url
 import io.ktor.http.headersOf
 import io.ktor.http.parseHeaderValue
-import io.ktor.http.takeFrom
 import io.ktor.http.withCharset
 import io.ktor.http.content.OutgoingContent
 import io.ktor.serialization.Configuration
@@ -55,9 +47,7 @@ import io.ktor.utils.io.ByteReadChannel
 import io.ktor.utils.io.ByteWriteChannel
 import io.ktor.utils.io.charsets.Charset
 
-import io.kjson.JSONCoPipeline
 import io.kjson.JSONConfig
-import io.kjson.JSONDeserializer
 import io.kjson.coStringifyJSON
 import io.kjson.toKType
 import io.kjson.ktor.io.KtorByteChannelCoAcceptor
@@ -65,8 +55,6 @@ import io.kjson.ktor.io.KtorByteChannelOutput
 import io.kjson.ktor.io.KtorOutgoingContent
 import io.kjson.util.CoOutputChannel
 import net.pwall.pipeline.IntCoAcceptor
-import net.pwall.pipeline.simpleCoAcceptor
-import net.pwall.pipeline.codec.CoDecoderFactory
 import net.pwall.pipeline.codec.CoEncoderFactory
 import net.pwall.util.CoOutput
 
@@ -114,7 +102,13 @@ fun ContentNegotiation.Config.kjson(
 val applicationJSON = ContentType.Application.Json
 
 /** The `application/json` content type as a string */
-val applicationJSONString = ContentType.Application.Json.toString()
+val applicationJSONString = applicationJSON.toString()
+
+/** The `application/jsonl` content type */
+val applicationJSONLines = ContentType("application", "jsonl")
+
+/** The `application/json` content type as a string */
+val applicationJSONLinesString = applicationJSONLines.toString()
 
 /**
  * Get a [Headers] containing a `Content-Type` of `application/json`.
@@ -122,10 +116,21 @@ val applicationJSONString = ContentType.Application.Json.toString()
 fun contentTypeJSON(): Headers = headersOf(HttpHeaders.ContentType, applicationJSONString)
 
 /**
+ * Get a [Headers] containing a `Content-Type` of `application/jsonl`.
+ */
+fun contentTypeJSONLines(): Headers = headersOf(HttpHeaders.ContentType, applicationJSONLinesString)
+
+/**
  * Get a [Headers] containing a `Content-Type` of `application/json` with the specified [Charset].
  */
 fun contentTypeJSON(charset: Charset): Headers =
         headersOf(HttpHeaders.ContentType, applicationJSON.withCharset(charset).toString())
+
+/**
+ * Get a [Headers] containing a `Content-Type` of `application/jsonl` with the specified [Charset].
+ */
+fun contentTypeJSONLines(charset: Charset): Headers =
+        headersOf(HttpHeaders.ContentType, applicationJSONLines.withCharset(charset).toString())
 
 /**
  * Create a [KtorByteChannelOutput] (convenience function).
@@ -155,6 +160,25 @@ fun createStreamedJSONContent(
 }
 
 /**
+ * Create an [OutgoingContent] to stream JSON Lines output.
+ */
+fun createStreamedJSONLinesContent(
+    value: JSONLinesOutput,
+    contentType: ContentType = ContentType.Application.Json,
+    charset: Charset = Charsets.UTF_8,
+    config: JSONConfig = JSONConfig.defaultConfig,
+): OutgoingContent = KtorOutgoingContent(contentType.withCharset(charset)) {
+    val output = CoOutputChannel(
+        downstream = CoEncoderFactory.getEncoder(
+            charset = charset,
+            downstream = KtorByteChannelCoAcceptor(this),
+        )
+    )
+    value.coOutput(config, output)
+    output.close()
+}
+
+/**
  * Respond to a call using streamed data.
  *
  * @param   contentType     the [ContentType]
@@ -177,167 +201,21 @@ suspend fun ApplicationCall.respondStream(
 }
 
 /**
- * Make a client call, receiving the response as a stream.  The response must be in the form of a JSON array, and each
- * array item will be deserialized and passed to the `consumer` function as it is received.
+ * Respond to a call using JSON Lines form, using a [Flow].
  *
- * @param   T               the type of the array item
- * @param   urlString       the URL as a [String]
- * @param   method          the HTTP method (default GET)
- * @param   body            the request body if required
- * @param   expectedStatus  the expected response status (default 200 OK)
- * @param   config          the [JSONConfig] to use when deserializing
- * @param   consumer        the consumer function (will be called with each array item)
+ * @param   flow        a [Flow] supplying the data.
  */
-suspend inline fun <reified T : Any> HttpClient.receiveStreamJSON(
-    urlString: String,
-    method: HttpMethod = HttpMethod.Get,
-    body: Any = EmptyContent,
-    headers: Headers = Headers.Empty,
-    expectedStatus: HttpStatusCode = HttpStatusCode.OK,
-    config: JSONConfig = JSONConfig.defaultConfig,
-    noinline consumer: suspend (T) -> Unit
-) {
-    receiveStreamJSON(typeOf<T>(), urlString, method, body, headers, expectedStatus, config, consumer)
+suspend fun ApplicationCall.respondLines(flow: Flow<Any?>) {
+    respond(JSONLinesFlowOutput(flow))
 }
 
 /**
- * Make a client call, receiving the response as a stream.  The response must be in the form of a JSON array, and each
- * array item will be deserialized and passed to the `consumer` function as it is received.
+ * Respond to a call using JSON Lines form, using a [ReceiveChannel].
  *
- * @param   T               the type of the array item
- * @param   type            the type of the array item as a [KType]
- * @param   urlString       the URL as a [String]
- * @param   method          the HTTP method (default GET)
- * @param   body            the request body if required
- * @param   expectedStatus  the expected response status (default 200 OK)
- * @param   config          the [JSONConfig] to use when deserializing
- * @param   consumer        the consumer function (will be called with each array item)
+ * @param   channel     a [ReceiveChannel] supplying the data.
  */
-suspend fun <T : Any> HttpClient.receiveStreamJSON(
-    type: KType,
-    urlString: String,
-    method: HttpMethod = HttpMethod.Get,
-    body: Any = EmptyContent,
-    headers: Headers = Headers.Empty,
-    expectedStatus: HttpStatusCode = HttpStatusCode.OK,
-    config: JSONConfig = JSONConfig.defaultConfig,
-    consumer: suspend (T) -> Unit
-) {
-    val requestBuilder = HttpRequestBuilder()
-    requestBuilder.url.takeFrom(urlString)
-    requestBuilder.method = method
-    requestBuilder.headers.appendAll(headers)
-    if (body !== EmptyContent) {
-        requestBuilder.setBody(body)
-        if (!requestBuilder.headers.contains(HttpHeaders.ContentType))
-            requestBuilder.headers[HttpHeaders.ContentType] = applicationJSONString
-    }
-    executeStreamJSON(type, requestBuilder, expectedStatus, config, consumer)
-}
-
-/**
- * Make a client call, receiving the response as a stream.  The response must be in the form of a JSON array, and each
- * array item will be deserialized and passed to the `consumer` function as it is received.
- *
- * @param   T               the type of the array item
- * @param   url             the URL as a [Url]
- * @param   method          the HTTP method (default GET)
- * @param   body            the request body if required
- * @param   expectedStatus  the expected response status (default 200 OK)
- * @param   config          the [JSONConfig] to use when deserializing
- * @param   consumer        the consumer function (will be called with each array item)
- */
-suspend inline fun <reified T : Any> HttpClient.receiveStreamJSON(
-    url: Url,
-    method: HttpMethod = HttpMethod.Get,
-    body: Any = EmptyContent,
-    headers: Headers = Headers.Empty,
-    expectedStatus: HttpStatusCode = HttpStatusCode.OK,
-    config: JSONConfig = JSONConfig.defaultConfig,
-    noinline consumer: suspend (T) -> Unit
-) {
-    receiveStreamJSON(typeOf<T>(), url, method, body, headers, expectedStatus, config, consumer)
-}
-
-/**
- * Make a client call, receiving the response as a stream.  The response must be in the form of a JSON array, and each
- * array item will be deserialized and passed to the `consumer` function as it is received.
- *
- * @param   T               the type of the array item
- * @param   type            the type of the array item as a [KType]
- * @param   url             the URL as a [Url]
- * @param   method          the HTTP method (default GET)
- * @param   body            the request body if required
- * @param   expectedStatus  the expected response status (default 200 OK)
- * @param   config          the [JSONConfig] to use when deserializing
- * @param   consumer        the consumer function (will be called with each array item)
- */
-suspend fun <T : Any> HttpClient.receiveStreamJSON(
-    type: KType,
-    url: Url,
-    method: HttpMethod = HttpMethod.Get,
-    body: Any = EmptyContent,
-    headers: Headers = Headers.Empty,
-    expectedStatus: HttpStatusCode = HttpStatusCode.OK,
-    config: JSONConfig = JSONConfig.defaultConfig,
-    consumer: suspend (T) -> Unit
-) {
-    val requestBuilder = HttpRequestBuilder()
-    requestBuilder.url.takeFrom(url)
-    requestBuilder.method = method
-    requestBuilder.headers.appendAll(headers)
-    if (body !== EmptyContent) {
-        requestBuilder.setBody(body)
-        if (!requestBuilder.headers.contains(HttpHeaders.ContentType))
-            requestBuilder.headers[HttpHeaders.ContentType] = applicationJSONString
-    }
-    executeStreamJSON(type, requestBuilder, expectedStatus, config, consumer)
-}
-
-/**
- * Make a client call with the parameters supplied in a [HttpRequestBuilder], receiving the response as a stream.  The
- * response must be in the form of a JSON array, and each array item will be deserialized and passed to the `consumer`
- * function as it is received.
- *
- * @param   T               the type of the array item
- * @param   type            the type of the array item as a [KType]
- * @param   expectedStatus  the expected response status (default 200 OK)
- * @param   config          the [JSONConfig] to use when deserializing
- * @param   consumer        the consumer function (will be called with each array item)
- */
-@Suppress("UNCHECKED_CAST")
-suspend fun <T : Any> HttpClient.executeStreamJSON(
-    type: KType,
-    requestBuilder: HttpRequestBuilder,
-    expectedStatus: HttpStatusCode = HttpStatusCode.OK,
-    config: JSONConfig = JSONConfig.defaultConfig,
-    consumer: suspend (T) -> Unit
-) {
-    prepareRequest(requestBuilder).execute { response ->
-        if (response.status == expectedStatus) {
-            val parsedContentTypeHeader = response.headers.parsedContentTypeHeader()
-            if (parsedContentTypeHeader?.value != applicationJSONString)
-                throw JSONKtorException("Content-Type not $applicationJSONString - ${requestBuilder.url}")
-            val charsetName = parsedContentTypeHeader.getParam("charset") ?: config.charset.name()
-            val pipeline = CoDecoderFactory.getDecoder(
-                charsetName = charsetName,
-                downstream = JSONCoPipeline(simpleCoAcceptor {
-                    val item = JSONDeserializer.deserialize(type, it, config) as T? ?:
-                            throw JSONKtorException("Streaming array item was null - ${requestBuilder.url}")
-                    consumer(item)
-                }),
-            )
-            response.body<ByteReadChannel>().copyToPipeline(pipeline, config.readBufferSize)
-        }
-        else
-            throw JSONKtorClientException(
-                urlString = requestBuilder.url.toString(),
-                statusCode = response.status,
-                responseHeaders = response.headers,
-                responseBody = response.body(),
-                config = config,
-            )
-    }
+suspend fun ApplicationCall.respondLines(channel: ReceiveChannel<Any?>) {
+    respond(JSONLinesChannelOutput(channel))
 }
 
 /**