+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.2] - 2023-07-19
+### Added
+- `JSONKtorStream`: functions extracted from `JSONKtorFunctions`
+- `JSONLinesOutput`, `JSONLinesFlowOutput`, `JSONLinesChannelOutput`: JSON Lines support
+### Changed
+- `JSONKtor`, `JSONKtorFunctions`: added support for JSON Lines
+- `pom.xml`: updated dependency version
+
 ## [1.1] - 2023-07-10
 ### Changed
 - `pom.xml`: updated version of kjson